feat: Add option to open diff in vertical split
This commit introduces a new option `g:silentsaver_diff_vertical` to control whether the diff is opened in a vertical split.

The default value is `v:false`.
When set to `v:true`, the diff will be opened in a vertical split.

This change provides more flexibility for users who prefer to view diffs in a vertical layout.